cap 3, team 3, fixed account dao test error

+++ 09-Capstone/student/java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDAO.java
@@ -6,7 +6,7 @@ import com.techelevator.tenmo.model.Account;
 
 public interface AccountDAO {
 	
-	public Account getAccountByUserId(int userId);
+	public Account getAccountByUserId(long userId);
 
 	public Account updateAccountBalance(String username, BigDecimal amount) throws OverdraftException;
